RTC and Push Button

+++ 326Final/ST7735.h
@@ -91,7 +91,7 @@
 #ifndef _ST7735H_
 #define _ST7735H_
 
-extern int LCDSelect;
+extern int8_t LCDSelect;
 // some flags for ST7735_InitR()
 enum initRFlags{
   none,
